Correct compatibility with Cygwin

+++ uninstall.sh
@@ -8,6 +8,9 @@ cd "`dirname "$0"`"
 
 if [ "`ls -al "$binDir" | fgrep " $binName " | sed 's/.*-> \(.*\)/\1/'`" = "$PWD/$srcFile" ]
 then
-	rm -rf "$binDir/$binName"
-	mv "$binDir/$binName.bak" "$binDir/$binName"
+	rm "$binDir/$binName"
+	if [ -e "$binDir/.$binName.bak" ]
+	then
+		mv "$binDir/$binName.bak" "$binDir/$binName"
+	fi
 fi
